# figabrain-social/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FIGABRAIN Social API
  version: 1.0.0
  description: >
    REST API for FIGABRAIN Social — a Web3-native social network where every
    account is simultaneously a social profile and an internal crypto
    wallet, and engagement is rewarded with Brain Points.
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        bio: { type: string }
        avatarUrl: { type: string, nullable: true }
        rank: { type: string, enum: [NPC, NORMIE, SIGMA, GIGACHAD, BRAIN_LORD, MEME_EMPEROR, GALAXY_BRAIN, NEURAL_GOD] }
        brainPoints: { type: number }
        xp: { type: integer }
        reputation: { type: integer }
        walletAddress: { type: string, nullable: true }
        walletBalance: { type: number }
        postsCount: { type: integer }
        followersCount: { type: integer }
        followingCount: { type: integer }
    Post:
      type: object
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        imageUrls: { type: array, items: { type: string } }
        linkUrl: { type: string, nullable: true }
        gifUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/User" }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        repostsCount: { type: integer }
        likedByMe: { type: boolean }
...
security:
  - bearerAuth: []
paths:
...
  /posts/feed:
    get:
      summary: Paginated chronological feed
      security: []
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: Feed page
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Post" }
                  nextCursor: { type: string, nullable: true }
  /posts:
    post:
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                imageUrls: { type: array, items: { type: string } }
                linkUrl: { type: string }
                gifUrl: { type: string }
      responses:
        "201": { description: Post created }
        "409": { description: Duplicate content detected }
...